Refactored IdfParser to remove code duplication. Added better handling of missing information in the file. Added validation to check if the number of pubmed IDs matches the number of publication titles. Moved Idf related clasees out of sdrf package.
