+++ base/src/main/java/uk/ac/ebi/atlas/experimentimport/idf/IdfParser.java
@@ -1,14 +1,15 @@
-package uk.ac.ebi.atlas.experimentimport.condensedSdrf;
+package uk.ac.ebi.atlas.experimentimport.idf;
 
 import com.google.common.collect.ImmutableSet;
 import org.apache.commons.lang.math.NumberUtils;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import uk.ac.ebi.atlas.commons.readers.TsvStreamer;
-import uk.ac.ebi.atlas.experimentimport.idf.IdfParserOutput;
 
 import javax.inject.Inject;
 import javax.inject.Named;
 import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -23,6 +24,8 @@ public class IdfParser {
     private static final String EXPECTED_CLUSTERS_ID = "Comment[ExpectedClusters]";
 
     private IdfStreamerFactory idfReaderFactory;
+
+    private  Map<String, List<String>> parsedIdf;
 
     @Inject
     public IdfParser(IdfStreamerFactory idfReaderFactory) {
@@ -57,7 +60,7 @@ public class IdfParser {
     public IdfParserOutput newParse(String experimentAccession) {
         try (TsvStreamer idfStreamer = idfReaderFactory.create(experimentAccession)) {
 
-            Map<String, List<String>> parsedIdf =
+            parsedIdf =
                     idfStreamer.get()
                             .filter(line -> line.length > 1)
                             .filter(line -> INVESTIGATION_TITLE_ID.equalsIgnoreCase(line[0].trim()) ||
@@ -72,13 +75,20 @@ public class IdfParser {
                                             .filter(item -> !item.isEmpty())
                                             .collect(Collectors.toList())));
 
-            String title = parsedIdf.getOrDefault(AE_EXPERIMENT_DISPLAY_NAME_ID.toUpperCase(), parsedIdf.get(INVESTIGATION_TITLE_ID.toUpperCase()))
+            String title = getParsedOutputByKey(AE_EXPERIMENT_DISPLAY_NAME_ID, getParsedOutputByKey(INVESTIGATION_TITLE_ID, Collections.emptyList()))
                     .stream()
                     .findFirst()
                     .orElse("");
 
-            Map<String, String> publicationDetails = getPublicationDetailsMap(parsedIdf.get(PUBMED_ID.toUpperCase()), parsedIdf.get(PUBLICATION_TITLE_ID.toUpperCase()));
-            int expectedClusters = NumberUtils.toInt(parsedIdf.get(EXPECTED_CLUSTERS_ID.toUpperCase()).stream().findFirst().orElse(""), 0);
+            Map<String, String> publicationDetails = getPublicationDetailsMap(
+                    getParsedOutputByKey(PUBMED_ID, Collections.emptyList()),
+                    getParsedOutputByKey(PUBLICATION_TITLE_ID, Collections.emptyList()));
+
+            int expectedClusters = NumberUtils.toInt(getParsedOutputByKey(EXPECTED_CLUSTERS_ID, Collections.emptyList())
+                        .stream()
+                        .findFirst()
+                        .orElse(""),
+                    0);
 
             return new IdfParserOutput(
                     title,
@@ -97,7 +107,20 @@ public class IdfParser {
                 publicationDetails.put(pubmedIds.get(i), publicationTitles.get(i));
             }
         }
+        else {
+            throw new IdfParserException("There is a mismatch between the number of PubMed IDs (" + pubmedIds.size() + ") and the number of publication titles " + publicationTitles.size() + ")");
+        }
         
         return  publicationDetails;
+    }
+
+    private List<String> getParsedOutputByKey(String key, List<String> outputIfEmpty) {
+        return parsedIdf.getOrDefault(key.toUpperCase(), outputIfEmpty);
+    }
+
+    class IdfParserException extends RuntimeException {
+        IdfParserException(String message) {
+            super(message);
+        }
     }
 }
